+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_amqp_transport_no_serializer.php
@@ -2,9 +2,7 @@
 
 $container->loadFromExtension('framework', array(
     'messenger' => array(
-        'serializer' => array(
-            'enabled' => false,
-        ),
+        'serializer' => false,
         'transports' => array(
             'default' => 'amqp://localhost/%2f/messages',
         ),