+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/messenger_transport_no_serializer.php
@@ -5,9 +5,6 @@ $container->loadFromExtension('framework', array(
         'enabled' => false,
     ),
     'messenger' => array(
-        'serializer' => array(
-            'enabled' => true,
-        ),
         'transports' => array(
             'default' => 'amqp://localhost/%2f/messages',
         ),