+++ src/Symfony/Component/Messenger/Tests/Transport/AmqpExt/AmqpSenderTest.php
@@ -16,7 +16,7 @@ use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Messenger\Transport\AmqpExt\AmqpSender;
 use Symfony\Component\Messenger\Transport\AmqpExt\Connection;
-use Symfony\Component\Messenger\Transport\Serialization\EncoderInterface;
+use Symfony\Component\Messenger\Transport\Serialization\SerializerInterface;
 
 /**
  * @requires extension amqp
@@ -28,13 +28,13 @@ class AmqpSenderTest extends TestCase
         $envelope = Envelope::wrap(new DummyMessage('Oy'));
         $encoded = array('body' => '...', 'headers' => array('type' => DummyMessage::class));
 
-        $encoder = $this->getMockBuilder(EncoderInterface::class)->getMock();
-        $encoder->method('encode')->with($envelope)->willReturnOnConsecutiveCalls($encoded);
+        $serializer = $this->getMockBuilder(SerializerInterface::class)->getMock();
+        $serializer->method('encode')->with($envelope)->willReturnOnConsecutiveCalls($encoded);
 
         $connection = $this->getMockBuilder(Connection::class)->disableOriginalConstructor()->getMock();
         $connection->expects($this->once())->method('publish')->with($encoded['body'], $encoded['headers']);
 
-        $sender = new AmqpSender($encoder, $connection);
+        $sender = new AmqpSender($serializer, $connection);
         $sender->send($envelope);
     }
 }